// widgets/ax-button/typings/AxButtonProps.d.ts
/**
 * This file was generated from AxButton.xml
 * WARNING: All changes made to this file will be overwritten
 * @author Mendix Widgets Framework Team
 */
import { CSSProperties } from "react";
import { ActionValue, DynamicValue } from "mendix";

export type VariantEnum = "contained" | "outlined" | "text";

export type ColorEnum = "primary" | "secondary" | "success" | "warning" | "error" | "info";

export type SizeEnum = "small" | "medium" | "large";

export interface AxButtonContainerProps {
    name: string;
    class: string;
    style?: CSSProperties;
    tabIndex?: number;
    label?: DynamicValue<string>;
    variant: VariantEnum;
    color: ColorEnum;
    size: SizeEnum;
    disabled: boolean;
    fullWidth: boolean;
    onClick?: ActionValue;
}

export interface AxButtonPreviewProps {
    /**
     * @deprecated Deprecated since version 9.18.0. Please use class property instead.
     */
    className: string;
    class: string;
    style: string;
    styleObject?: CSSProperties;
    readOnly: boolean;
    renderMode: "design" | "xray" | "structure";
    translate: (text: string) => string;
    label: string;
    variant: VariantEnum;
    color: ColorEnum;
    size: SizeEnum;
    disabled: boolean;
    fullWidth: boolean;
    onClick: {} | null;
}